add package duke
