+++ src/main/java/duke/Deadline.java
@@ -1,8 +1,10 @@
+package duke;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.time.temporal.ChronoUnit;
 
-public class Deadline extends Task{
+public class Deadline extends Task {
     protected LocalDate by;
 
     public Deadline(String description, String by) {
