+++ src/main/java/duke/Event.java
@@ -1,7 +1,8 @@
+package duke;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
-import java.time.temporal.ChronoUnit;
 
 public class Event extends Task{
     protected LocalDate at;
