+++ src/main/java/duke/TaskList.java
@@ -1,3 +1,5 @@
+package duke;
+
 import java.util.ArrayList;
 public class TaskList {
     ArrayList<Task> tasks;
