+++ src/main/java/duke/ToDo.java
@@ -1,3 +1,5 @@
+package duke;
+
 public class ToDo extends Task{
 
     public ToDo(String description){
